refactor: move artwirk miniature to entities layer
