+++ src/entities/painting/index.ts
@@ -1,1 +1,2 @@
 export type { Painting } from "./model/Painting";
+export { ArtworkMiniature } from "./ui/ArtworkMiniature";
