+++ src/entities/painting/ui/ArtworkMiniature.module.css
